Move CollectorRegistry to its own module
