+++ prometheus_client/core.py
@@ -2,7 +2,6 @@
 
 from __future__ import unicode_literals
 
-import copy
 import math
 import os
 import sys
@@ -11,6 +10,7 @@ import types
 
 from threading import Lock
 
+from .registry import REGISTRY
 from .metrics import *  # noqa
 from .metrics import _METRIC_LABEL_NAME_RE, _METRIC_NAME_RE, _RESERVED_METRIC_LABEL_NAME_RE
 from .context_managers import _ExceptionCounter, _InprogressTracker, _Timer
@@ -20,125 +20,6 @@ if sys.version_info > (3,):
 
 _INF = float("inf")
 _MINUS_INF = float("-inf")
-
-
-class CollectorRegistry(object):
-    '''Metric collector registry.
-
-    Collectors must have a no-argument method 'collect' that returns a list of
-    Metric objects. The returned metrics should be consistent with the Prometheus
-    exposition formats.
-    '''
-
-    def __init__(self, auto_describe=False):
-        self._collector_to_names = {}
-        self._names_to_collectors = {}
-        self._auto_describe = auto_describe
-        self._lock = Lock()
-
-    def register(self, collector):
-        '''Add a collector to the registry.'''
-        with self._lock:
-            names = self._get_names(collector)
-            duplicates = set(self._names_to_collectors).intersection(names)
-            if duplicates:
-                raise ValueError(
-                    'Duplicated timeseries in CollectorRegistry: {0}'.format(
-                        duplicates))
-            for name in names:
-                self._names_to_collectors[name] = collector
-            self._collector_to_names[collector] = names
-
-    def unregister(self, collector):
-        '''Remove a collector from the registry.'''
-        with self._lock:
-            for name in self._collector_to_names[collector]:
-                del self._names_to_collectors[name]
-            del self._collector_to_names[collector]
-
-    def _get_names(self, collector):
-        '''Get names of timeseries the collector produces.'''
-        desc_func = None
-        # If there's a describe function, use it.
-        try:
-            desc_func = collector.describe
-        except AttributeError:
-            pass
-        # Otherwise, if auto describe is enabled use the collect function.
-        if not desc_func and self._auto_describe:
-            desc_func = collector.collect
-
-        if not desc_func:
-            return []
-
-        result = []
-        type_suffixes = {
-            'counter': ['_total', '_created'],
-            'summary': ['', '_sum', '_count', '_created'],
-            'histogram': ['_bucket', '_sum', '_count', '_created'],
-            'info': ['_info'],
-        }
-        for metric in desc_func():
-            for suffix in type_suffixes.get(metric.type, ['']):
-                result.append(metric.name + suffix)
-        return result
-
-    def collect(self):
-        '''Yields metrics from the collectors in the registry.'''
-        collectors = None
-        with self._lock:
-            collectors = copy.copy(self._collector_to_names)
-        for collector in collectors:
-            for metric in collector.collect():
-                yield metric
-
-    def restricted_registry(self, names):
-        '''Returns object that only collects some metrics.
-
-        Returns an object which upon collect() will return
-        only samples with the given names.
-
-        Intended usage is:
-            generate_latest(REGISTRY.restricted_registry(['a_timeseries']))
-
-        Experimental.'''
-        names = set(names)
-        collectors = set()
-        with self._lock:
-            for name in names:
-                if name in self._names_to_collectors:
-                    collectors.add(self._names_to_collectors[name])
-        metrics = []
-        for collector in collectors:
-            for metric in collector.collect():
-                samples = [s for s in metric.samples if s[0] in names]
-                if samples:
-                    m = Metric(metric.name, metric.documentation, metric.type)
-                    m.samples = samples
-                    metrics.append(m)
-
-        class RestrictedRegistry(object):
-            def collect(self):
-                return metrics
-
-        return RestrictedRegistry()
-
-    def get_sample_value(self, name, labels=None):
-        '''Returns the sample value, or None if not found.
-
-        This is inefficient, and intended only for use in unittests.
-        '''
-        if labels is None:
-            labels = {}
-        for metric in self.collect():
-            for s in metric.samples:
-                if s.name == name and s.labels == labels:
-                    return s.value
-        return None
-
-
-REGISTRY = CollectorRegistry(auto_describe=True)
-'''The default registry.'''
